Refactor ConfigFactoryDocumentTest to Shared and JVM - JS and Native tests from 318 to 346
