+++ sconfig/shared/src/test/scala/org/ekrich/config/impl/PathTest.scala
@@ -8,7 +8,7 @@ import org.junit._
 import scala.jdk.CollectionConverters._
 import org.ekrich.config.ConfigException
 
-class PathTest extends TestUtils {
+class PathTest extends TestUtilsShared {
   @Test
   def pathEquality(): Unit = {
     // note: foo.bar is a single key here
